# Conflicts:
#	src/js/components/FlowListContainer/index.jsx
#	src/js/components/shared/Inspector/BaseInspector/index.jsx

+++ kskp/api.py
@@ -4,7 +4,8 @@ from pathlib import Path
 from .engine.data3 import *
 from flask import Blueprint, request, session, jsonify, send_from_directory, render_template
 from .auth import login_required_api
-from .navigation import update_navigation
+from .utils.navigation import update_navigation
+from .utils.api_base import api_base
 from .model import (
     start_project,
     get_projects_by_user_id,
@@ -23,9 +24,20 @@ from .model import (
     update_user_by_id,
     write_data_to_json,
     make_flow_path,
-    copy_flow_by_uuid
+    copy_flow_by_uuid,
+    get_all_frame_uuid_in_frame,
+    get_frame_dir_path,
+    get_cache_dir_path
 )
-from .activity import (
+from .library import (
+    # data3.pyのFrameクラスと名称を被らないようにAS別名を付ける
+    # (将来的にdata3.pyのFrameと統合したい)
+    Frame as FrameModel,
+    Folder,
+    FRAME_FOLDER_UUID,
+    CACHE_FOLDER_UUID
+)
+from .utils.activity import (
     make_unfinished_history,
     make_finished_history
 )
@@ -37,7 +49,6 @@ api = Blueprint('api', __name__)
 DATAFRAME_DIR_PATH = api.root_path / Path('data/frames')
 JOBS_DIR_PATH = api.root_path / Path('data/jobs')
 FLOWS_DIR_PATH = api.root_path / Path('data/flows')
-
 @api.route('/projects', methods=['POST'])
 @login_required_api
 def new_project():
@@ -199,29 +210,38 @@ def execute_flow_by_add_inputs(request):
 
     upload_file_list = []
 
+    def get_frame_obj(frame_uuid):
+        """
+        指定したフレームからエンジンのFrameオブジェクトを作成して返す
+        """
+        # フレームを置き換える
+        frame = FrameModel.find_by_uuid(frame_uuid)
+        if frame is None:
+            # ライブラリにフレームが無い場合は従来のフォルダ内を探す
+            source = PathFileSource('csv', DATAFRAME_DIR_PATH , frame_uuid + '.csv')
+        else:
+            # ライブラリにフレームが存在する場合はライブラリから取得する
+            source = PathFileSource('csv', Path(api.root_path).parent / frame.path_obj.parent, frame.path_obj.name)
+        return Frame(str(uuid.uuid4()), source)
+        
     for port in flow_json['ports'][0]:
-        frame_uuid = ''
-
         # frame（既にkskpに存在するデータソース）の場合
         if request.form.get(port['name']) is not None:
             # フレームを置き換える
             frame_uuid = request.form.get(port['name'])
-            inputs[port['name']] = Frame(str(uuid.uuid4()), PathFileSource('csv', DATAFRAME_DIR_PATH , frame_uuid + '.csv'))
-            continue
+            inputs[port['name']] = get_frame_obj(frame_uuid)
 
         # 新たにkskpにアップロードする場合
         file = request.files.get(port['name'])
         if file is not None:
             # ファイルアップロードして、フレームを置き換える
             frame_uuid = upload_frame(file, '')['uuid']
-            inputs[port['name']] = Frame(str(uuid.uuid4()), PathFileSource('csv', DATAFRAME_DIR_PATH , frame_uuid + '.csv'))
-
+            inputs[port['name']] = get_frame_obj(frame_uuid)
             # 使うかわからないけど、uploadしたファイルを覚えておく
             upload_file_list.append(frame_uuid)
-            continue
 
     # フローの実行
-    result = execute_flow(flow_uuid, None, False, None, inputs, args)
+    result = execute_flow(flow_uuid, None, False, None, inputs, args, flow_label=flow_json['label'])
 
     return result
 
@@ -408,6 +428,7 @@ def fetch_visualizers():
     return jsonify({'success': True, 'data': visualizers})
 
 @api.route('/frames', methods=['GET', 'POST'])
+@login_required_api
 def make_new_frame():
     """
     新しいframeを作成する
@@ -417,9 +438,27 @@ def make_new_frame():
     no_contents = False
 
     if 'file' in request.files:
-        # ファイルがPOSTで送信されてきたらアップロードだとみなす
-        frame = upload_frame(request.files.get('file'), request.form.get('file_name'))
-        return jsonify({'success': True, "data": frame})
+        if 'parent' in request.form and 'label' in request.form:
+            try:
+                # parentとlabel属性があれば新形式のPOST /framesだとみなす
+                new_frame = FrameModel(request.form.get('parent')
+                                     , request.form.get('label')
+                                     , request.files.get('file').stream
+                                     , creator=session['user_id']
+                                     , modifier=session['user_id'])
+                # documentレコードをDBに格納する
+                new_frame.save()
+                return jsonify({'success': True, 'data': new_frame.to_json()})
+            except Exception as e:
+                return jsonify({
+                                'success': False,
+                                'code'   : -1,
+                                'message': str(e)
+                                })
+        else:
+            # ファイルがPOSTで送信されてきたらアップロードだとみなす
+            frame = upload_frame(request.files.get('file'), request.form.get('file_name'))
+            return jsonify({'success': True, "data": frame})
     elif 'from' in request.args:
         if '.' in request.args['from']:
             # ドットで区切って、具体的に一つだけstepを指定することができる
@@ -436,7 +475,9 @@ def make_new_frame():
 
         limit = int(request.args.get('limit')) if request.args.get('limit') else None
 
-        result = execute_flow(flow_uuid, step_paths=step_id, no_contents=no_contents, limit=limit)
+        flow_json = fetch_flow_by_uuid(flow_uuid)
+
+        result = execute_flow(flow_uuid, step_paths=step_id, no_contents=no_contents, limit=limit, flow_label=flow_json['label'])
 
         return result
     elif request.form.get('flow_uuid'):
@@ -452,6 +493,8 @@ import os
 import time
 
 @api.route('/frames/<frame_uuid>')
+@login_required_api
+@api_base
 def fetch_frame(frame_uuid):
     """
     指定したframeを直接UUIDで指定して取得する
@@ -461,7 +504,17 @@ def fetch_frame(frame_uuid):
     limit = int(request.args.get('limit')) if request.args.get('limit') else None
     no_contents = True if request.args.get('no_contents') else False
 
-    file_path = DATAFRAME_DIR_PATH / Path('%s.csv' % frame_uuid)
+    frame = FrameModel.find_by_uuid(frame_uuid)
+
+    if frame is None:
+        # ライブラリにフレームが無い場合は従来のフォルダ内を探す
+        file_path = DATAFRAME_DIR_PATH / Path('%s.csv' % frame_uuid)
+    else:
+        # ライブラリにフレームが存在する場合はライブラリから取得する
+        limit = 999 if limit is None else limit
+        no_contents = request.args.get('no_contents') is not None
+        file_path = frame.path_obj
+
     result = csv_to_frame(file_path, no_contents=no_contents, offset=offset, limit=limit)
 
     if request.args.get('header_only') == '1':
@@ -471,8 +524,42 @@ def fetch_frame(frame_uuid):
             headers.append(column.replace('\n',''))
         result = headers
 
+    return result
 
-    return jsonify({'success': True, 'data': result})
+@api.route('/frames/<frame_uuid>', methods=['PUT'])
+@login_required_api
+@api_base
+def update_frame(frame_uuid):
+    """
+    指定したframeのラベル名を変更する
+    """
+    label = request.json['label']
+    modifier = session['user_id']
+    return FrameModel.update_data(frame_uuid, label, modifier)
+
+@api.route('/frames/<frame_uuid>', methods=['DELETE'])
+@login_required_api
+@api_base
+def delete_frame(frame_uuid):
+    """
+    指定したframeを物理削除する
+    """
+    frame = FrameModel.find_by_uuid(frame_uuid)
+    if frame is None:
+        raise Exception('no frame exists.')
+    
+    # 削除しようとするframeが、フローで使用されている場合は例外を送出する
+    for flow_path in Path(app.config['FLOW_PATH']).iterdir():
+        if not flow_path.suffix == '.json':
+            continue
+        flow_uuid = flow_path.stem
+        using_frame_uuids = get_all_frame_uuid_in_frame(flow_uuid)
+        if frame_uuid in using_frame_uuids:
+            raise Exception('このCSVファイルはフロー(%s)で使用しているため削除できません' % flow_uuid)
+
+    # フレームを削除する
+    frame.delete()
+    return frame
 
 def csv_to_frame(file_path, no_contents=False, offset=0, limit=None):
     """
@@ -524,23 +611,37 @@ def download_file():
     # JST = timezone(timedelta(hours=+9), 'JST')
     # date = datetime.now(JST)
 
-    # ダウンロードファイルの名前
-    if frame_uuid == 'テスト':
-        downloadFileName = frame_uuid  + '.' + ext
+
+    frame = FrameModel.find_by_uuid(frame_uuid)
+    file_path = frame.path if frame is not None else None
+
+    if file_path is None:
+        # ライブラリにフレームが無い場合は従来のフォルダ内を探す
+
+        # ダウンロードファイルの名前
+        if frame_uuid == 'テスト':
+            downloadFileName = frame_uuid  + '.' + ext
+        else:
+            downloadFileName = label + '.' + ext
+
+        # ダウンロード対象のファイルの名前
+        downloadFile = frame_uuid + '_sjis.' + ext
+        sjis_path = DATAFRAME_DIR_PATH / downloadFile
+        # sjis版がなかったらutf8版を落とす（今の所sjis版はオムロンさま専用なので）
+        if not sjis_path.exists():
+            downloadFile = frame_uuid + '.' + ext
+
+        return send_from_directory(DATAFRAME_DIR_PATH, downloadFile, as_attachment = True,
+                                   attachment_filename = downloadFileName, mimetype = 'text/csv')
     else:
-        downloadFileName = label + '.' + ext
+        # ライブラリにフレームが存在する場合はライブラリから取得する
+        dir_path = Path(api.root_path).parent / Path(os.path.dirname(file_path))
+        file_name = os.path.basename(file_path)
+        return send_from_directory(dir_path, file_name, as_attachment = True,
+                                   attachment_filename = file_name, mimetype = 'text/csv')
+        
 
-    # ダウンロード対象のファイルの名前
-    downloadFile = frame_uuid + '_sjis.' + ext
-    sjis_path = DATAFRAME_DIR_PATH / downloadFile
-    # sjis版がなかったらutf8版を落とす（今の所sjis版はオムロンさま専用なので）
-    if not sjis_path.exists():
-        downloadFile = frame_uuid + '.' + ext
-
-    return send_from_directory(DATAFRAME_DIR_PATH, downloadFile, as_attachment = True,
-                               attachment_filename = downloadFileName, mimetype = 'text/csv')
-
-def execute_flow(flow_uuid, step_paths, no_contents, limit=None, inputs={}, args={}):
+def execute_flow(flow_uuid, step_paths, no_contents, limit=None, inputs={}, args={}, flow_label=None):
 
     # 指定されたIDのフローが存在するかどうかをチェックする
     # まずは、フローファイル一覧を取得する
@@ -555,7 +656,7 @@ def execute_flow(flow_uuid, step_paths, no_contents, limit=None, inputs={}, args
                         })
     else:
         try:
-            result_data, caches_data = execute_flow_internal(flow_uuid, step_paths, no_contents, limit, inputs, args)
+            result_data, caches_data = execute_flow_internal(flow_uuid, step_paths, no_contents, limit, inputs, args, flow_label=flow_label)
             if not result_data:
                 return jsonify({
                                     'success': False,
@@ -570,7 +671,7 @@ def execute_flow(flow_uuid, step_paths, no_contents, limit=None, inputs={}, args
             return jsonify({
                                 'success': False,
                                 'code': -1,
-                                'message': repr(e)
+                                'message': str(e)
                             })
 
 @api.route('/jobs', methods=['GET'])
@@ -584,6 +685,8 @@ def jobs():
 
     # jobsリストの作成
     for job_path in Path(JOBS_DIR_PATH).iterdir():
+        if not job_path.suffix == '.json':
+            continue
         data = json.loads(job_path.read_text(encoding='utf-8'))
         if 'flow' in request.args:
             if data['flow']['uuid'] == request.args['flow']:
@@ -675,8 +778,6 @@ def delete_cache():
     flow_uuid = ofs[0]
     datum_id = ofs[1]
 
-    frame_name = DATAFRAME_DIR_PATH / ('caches_' + flow_uuid + '_' + datum_id + '.csv')
-
     p = FLOWS_DIR_PATH.joinpath(flow_uuid + '.json')
     j = json.loads(p.read_text(), encoding='utf-8')
 
@@ -687,12 +788,12 @@ def delete_cache():
             j['nodes'][i]['cacheCreatedAt'] = None
 
             # キャッシュを削除する（増え続けると困るので）
-            frame_path = DATAFRAME_DIR_PATH / (frame_uuid + '.csv')
-            if frame_path.exists():
-                frame_path.unlink()
-            sjis_path = DATAFRAME_DIR_PATH / (frame_uuid + '_sjis.csv')
-            if sjis_path.exists():
-                sjis_path.unlink()
+            frame = FrameModel.find_by_uuid(frame_uuid)
+            if frame is not None:
+                sjis_path = frame.path_obj.parent / (frame_uuid + '_sjis.csv')
+                if sjis_path.exists():
+                    sjis_path.unlink()
+                frame.delete()
 
     update_flow_by_uuid(p.stem, j)
 
@@ -829,7 +930,7 @@ def execute_fifo():
 
     t2 = time.time()
 
-    return jsonify({'success': True, 'speed': repr(t2 - t1)})
+    return jsonify({'success': True, 'speed': str(t2 - t1)})
 
 @api.route('/execute-direct')
 def execute_direct():
@@ -876,7 +977,7 @@ def execute_direct():
 
     t2 = time.time()
 
-    return jsonify({'success': True, 'speed': repr(t2 - t1)})
+    return jsonify({'success': True, 'speed': str(t2 - t1)})
 
 @api.route('/execute-direct2')
 def execute_direct2():
@@ -925,7 +1026,7 @@ def execute_direct3():
 
     return jsonify({'success': True, 'data': 'execute-direct3'})
 
-def execute_flow_internal(flow_uuid, step_paths=None, no_contents=False, limit=None, inputs={}, args={}):
+def execute_flow_internal(flow_uuid, step_paths=None, no_contents=False, limit=None, inputs={}, args={}, flow_label=None):
     """
     指定されたファイル名を元にフローファイルを取得して、
     その結果をパースしてDataFrameの形にして返す
@@ -933,16 +1034,79 @@ def execute_flow_internal(flow_uuid, step_paths=None, no_contents=False, limit=N
 
     now = datetime.now()
 
+
+    # フローの実行結果を格納するディレクトリパスを取得する
+    frame_folder_path_obj = get_frame_dir_path(session['user_id']).path_obj
+
+    # キャッシュフォルダを作成する
+    get_cache_dir_path(session['user_id'])
+
     @make_unfinished_history(now, session)
     @make_finished_history(now)
     def execute_flow_by_uuid(flow_uuid, inputs={}, args={}):
         from . import engine as e
         # data_path = (DATAFRAME_DIR_PATH / 'data').as_posix()
         with open(FLOWS_DIR_PATH.joinpath(f'{flow_uuid}.json'), 'r') as f:
-            return e.execute(flow_uuid, f.read(), step_paths=step_paths, frames_path=DATAFRAME_DIR_PATH.as_posix(), flows_path=FLOWS_DIR_PATH.as_posix(), inputs=inputs, arguments=args)
-
+            return e.execute(flow_uuid, f.read(), step_paths=step_paths, frames_path=frame_folder_path_obj.as_posix(), flows_path=FLOWS_DIR_PATH.as_posix(), inputs=inputs, arguments=args)
     result = execute_flow_by_uuid(flow_uuid=flow_uuid, inputs=inputs, args=args)
     nodes_dict = get_flow_nodes_by_uuid(flow_uuid)
+
+    def register_file_to_library(folder_uuid, uuid, label):
+        """
+        execute_flow_by_uuid()が出力した結果ファイルとキャッシュファイルをライブラリに登録する
+        ファイル名に用いているUUIDをライブラリエントリにも付番し、ライブラリに登録する
+        """
+        file_path_obj = Path.joinpath(frame_folder_path_obj, uuid + '.csv')
+        if Path(file_path_obj).is_file():
+            # 既に登録済みのuuidであれば登録処理をしない
+            # (lastsノードがキャッシュを出力する場合は、result['outputs]とresult['caches']の両方に同じファイル名が格納される)
+            if FrameModel.exists(uuid):
+                return
+            new_frame = FrameModel(folder_uuid,
+                                   label,
+                                   None,
+                                   creator=session['user_id'],
+                                   modifier=session['user_id'])
+            # フレームのuuidはエンジン内で付番されたUUIDとする
+            new_frame.uuid = uuid
+            new_frame.add_entry_from_path(file_path_obj.as_posix())
+
+    def register_sjis_file_to_library(folder_uuid, uuid, label):
+        """
+        execute_flow_by_uuid()が出力した_sjisファイルをライブラリに登録する
+        ファイル名にはUTF-8版と同じUUIDが_sjis版に使用されているため、ライブラリエントリには新たなUUIDを採番する
+        """
+        file_path_obj = Path.joinpath(frame_folder_path_obj, uuid + '_sjis.csv')
+        if Path(file_path_obj).is_file():
+            new_frame = FrameModel(folder_uuid,
+                                   label,
+                                   None,
+                                   creator=session['user_id'],
+                                   modifier=session['user_id'])
+            new_frame.add_entry_from_path(file_path_obj.as_posix())
+
+    # 出力されたデータフレームをライブラリに登録する
+    for key, value in result['outputs'].items():
+        if nodes_dict.get(key).get('label') is None:
+            # データフレームにlabel属性が定義されていない場合に備える
+            label = flow_label + '_' + nodes_dict.get(key).get('id')
+        else:
+            label = flow_label + '_' + nodes_dict.get(key).get('label')
+        register_file_to_library(FRAME_FOLDER_UUID, value.uuid, label)
+
+        # 出力された_sjisファイルをライブラリに登録する
+        register_sjis_file_to_library(FRAME_FOLDER_UUID, value.uuid, label + '_sjis')
+
+    # 出力されたキャッシュファイルをライブラリに登録する
+    for key, uuid in result['caches'].items():
+        if key is None or key.split('.')[1] is None:
+            label = flow_label + '_cache'
+        else:
+            label = flow_label + '_' + key.split('.')[1] + '_cache'
+        register_file_to_library(CACHE_FOLDER_UUID, uuid, label)
+
+        # 出力された_sjisファイルをライブラリに登録する
+        register_sjis_file_to_library(CACHE_FOLDER_UUID, uuid, label + '_sjis')
 
     # 結果の処理
     if no_contents:
@@ -1068,7 +1232,19 @@ def visualizer():
     ### ここから
     # ここから
     new_inputs = {}
-    new_inputs['i'] = Frame(str(uuid.uuid4()), PathFileSource('csv', DATAFRAME_DIR_PATH , request.json.get('inputs')['i'] + '.csv'))
+
+    frame_uuid = request.json.get('inputs')['i']
+    frame = FrameModel.find_by_uuid(frame_uuid)
+    
+    if frame is None:
+        # ライブラリにフレームが無い場合は従来のフォルダ内を探す
+        source = PathFileSource('csv', DATAFRAME_DIR_PATH, request.json.get('inputs')['i'] + '.csv')
+    else:
+        # ライブラリにフレームが存在する場合はライブラリから取得する
+        source = PathFileSource('csv', Path(api.root_path).parent / frame.path_obj.parent, frame.path_obj.name)
+
+    new_inputs['i'] = Frame(str(uuid.uuid4()), source)
+
     command = internal_commands.get(request.args.get('from'))
     # 残りの２つの引数はsrcsとdsts
     new_step = Step(command, request.json.get('args'), {}, {})
@@ -1081,7 +1257,9 @@ def visualizer():
 
     # テーブルコマンド
     if request.args.get('from') == 'csvtohtmltable':
+        print('RESULT', result)
         return render_template("visualize_table.html", header=result['header'], reader=result['reader'])
 
     # bokehのコマンド
     return render_template("visualize_component.html", script=result['script'], div=result['div'])
+
